+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
 v0.94.1
 
 - Add `grpc.BackupsAnonymousGetUploadForm` remote config, for both backup and backup media uploads. This is separate from the `grpc.AttachmentsGetUploadForm` config added previously, which applies to regular attachment uploads.
+
+- keytrans: Add reset account data field functionality for all platforms.